feat(profile): add order management components and user profile settings
- Implemented OrderMobileCard for mobile view of orders.
- Created OrderStatusProgress to visually represent order status.
- Developed OrderTable for desktop view of orders with details and cancel option.
- Added OrdersTab to encapsulate order history and display both mobile and desktop views.
- Introduced ProfileTab for displaying and updating user information.
- Created SettingsTab for account settings including password change.
- Refactored UserAccount page to utilize new components for better structure and readability.
- Added utility functions for formatting currency, dates, and generating IDs.
- Updated mock data to include slugs for cosmetics. #19
